(ol/stf): Polish code

+++ crates/ol/chain-types/src/common.rs
@@ -2,7 +2,7 @@ use strata_primitives::buf::Buf32;
 
 /// Type aliases for clarity
 pub type Slot = u64;
-pub type Epoch = u32;
+pub type Epoch = u64;
 pub type OLBlockId = Buf32;
 pub type L1BlockId = Buf32;
